feat(bloom): enable configuring material emissive intensity separately from global bloom
