+++ src/components/Background/Color/Color.component.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from 'react';
+
+export type BackgroundColorProps = {
+  /**
+   * CSS-style string. For example: '#000' or 'red'.
+   */
+  color: string;
+};
+
+export const BackgroundColor: FC<BackgroundColorProps> = ({ color }) => <color attach="background" args={[color]} />;
