+++ src/components/Background/index.ts
@@ -1,0 +1,2 @@
+export { BackgroundColor } from './Color/Color.component';
+export { Box } from './Box/Box.component';
