+++ src/components/BaseCanvas/BaseCanvas.component.tsx
@@ -27,7 +27,7 @@ export const BaseCanvas: FC<BaseCanvasProps> = ({
     dpr={dpr}
     camera={{ fov, position }}
     resize={{ scroll: true, debounce: { scroll: 50, resize: 0 } }}
-    style={{ ...style }}
+    style={{ ...style, background: 'transparent' }}
   >
     {children}
   </Canvas>
